coinbase/requirements.txt
coinbase-advanced-py>=1.8.2
pandas>=2.2
numpy>=1.26
pydantic>=2.7
python-dotenv>=1.0
httpx>=0.27
neo4j>=5.22
streamlit>=1.37
